pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enterprise-rag-system"
version = "0.1.0"
description = "Enterprise multi-tenant RAG platform monorepo scaffold."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "beautifulsoup4>=4.12,<5.0",
  "fastapi>=0.115,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyJWT>=2.10,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "pydantic>=2.9,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "redis>=5.2,<6.0",
  "sqlalchemy>=2.0.36,<3.0",
  "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "mypy>=1.13,<2.0",
  "pre-commit>=4.0,<5.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.8,<1.0",
]

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = [
  "apps/api/src",
  "apps/worker/src",
  "apps/indexer/src",
  "packages/core/src",
  "packages/auth/src",
  "packages/ingestion/src",
  "packages/retrieval/src",
  "packages/llm/src",
  "packages/observability/src",
  "packages/evals/src",
]
include = ["enterprise_rag*", "app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = [
  "apps/api/src",
  "apps/worker/src",
  "apps/indexer/src",
  "packages/core/src",
  "packages/auth/src",
  "packages/ingestion/src",
  "packages/retrieval/src",
  "packages/llm/src",
  "packages/observability/src",
  "packages/evals/src",
]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [
  "apps/api/src",
  "apps/worker/src",
  "apps/indexer/src",
  "packages/core/src",
  "packages/auth/src",
  "packages/ingestion/src",
  "packages/retrieval/src",
  "packages/llm/src",
  "packages/observability/src",
  "packages/evals/src",
]
